# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "batsim",
    "description": "Residential battery fleet simulator: physics-faithful virtual homes behind an OpenAPI-first HTTP API for dispatch-strategy testing (ERCOT).",
    "license": {
      "name": "MIT OR Apache-2.0",
      "identifier": "MIT OR Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/v1/dispatch": {
      "post": {
        "tags": [
          "dispatch"
        ],
        "summary": "Send a command to a home/fleet subset.",
        "operationId": "dispatch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DispatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Command accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DispatchResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Unknown target",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Idempotency conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Empty target set",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/dispatch/commands/{command_id}": {
      "get": {
        "tags": [
          "dispatch"
        ],
        "summary": "Command status and per-target execution detail.",
        "operationId": "get_command",
        "parameters": [
          {
            "name": "command_id",
            "in": "path",
            "description": "Command id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The command",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandDoc"
                }
              }
            }
          },
          "404": {
            "description": "Unknown command",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "dispatch"
        ],
        "summary": "Cancel a command's still-queued targets.",
        "operationId": "cancel_command",
        "parameters": [
          {
            "name": "command_id",
            "in": "path",
            "description": "Command id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Command cancelled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkDoc"
                }
              }
            }
          },
          "404": {
            "description": "Unknown command",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Command already finished",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/fleets": {
      "get": {
        "tags": [
          "fleets"
        ],
        "summary": "List fleets.",
        "operationId": "list_fleets",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Page size (1..=1000, default 100).",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "maximum": 1000,
              "minimum": 1
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Opaque continuation cursor from a previous page.",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(?:[A-Za-z0-9_-]{4})*(?:[A-Za-z0-9_-]{2,3})?$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Page of fleets",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FleetsPage"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "fleets"
        ],
        "summary": "Create a fleet from a manifest.",
        "operationId": "create_fleet",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FleetManifest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Fleet created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FleetDoc"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Idempotency conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Composition rule violation",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/fleets/{id}": {
      "get": {
        "tags": [
          "fleets"
        ],
        "summary": "Get one fleet.",
        "operationId": "get_fleet",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Fleet id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The fleet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FleetDoc"
                }
              }
            }
          },
          "404": {
            "description": "Unknown fleet",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "fleets"
        ],
        "summary": "Delete a fleet and its homes.",
        "operationId": "delete_fleet",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Fleet id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Fleet deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkDoc"
                }
              }
            }
          },
          "404": {
            "description": "Unknown fleet",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Simulation is running",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/fleets/{id}:dispatch": {
      "post": {
        "tags": [
          "fleets"
        ],
        "summary": "Dispatch to every home in a fleet.",
        "operationId": "dispatch_fleet",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Fleet id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FleetDispatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Command accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DispatchResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Unknown fleet",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Idempotency conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DispatchRequest": {
        "type": "object",
        "description": "Dispatch request.",
        "required": [
          "target",
          "action"
        ],
        "properties": {
          "action": {
            "$ref": "#/components/schemas/ActionSpec",
            "description": "Action."
          },
          "command_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Client-supplied command id (ULID/uuid). Retries with the same id\nare deduplicated before enqueue. Server-assigned when omitted."
          },
          "execution": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ExecutionSpec",
                "description": "Execution shaping."
              }
            ]
          },
          "target": {
            "$ref": "#/components/schemas/TargetSpec",
            "description": "Target set."
          }
        },
        "additionalProperties": false
      },
...
      "OkDoc": {
        "type": "object",
        "description": "Empty success body for action endpoints with nothing to return.",
        "required": [
          "ok"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "description": "Always true."
          }
        }
      },
...
    }
  },
...
}